requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=20.1.0=py39h27cfd23_1
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py39hdd57654_1
brotlipy=0.7.0=py39h27cfd23_1003
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py39hf3d152e_1
cffi=1.14.6=py39h400218f_0
chardet=4.0.0=py39h06a4308_1003
conda=4.11.0=py39hf3d152e_0
conda-package-handling=1.7.3=py39h27cfd23_1
cryptography=3.4.7=py39hd23ed53_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py39h06a4308_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
glib=2.69.1=h5202010_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=4.8.2=py39h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.4.1=py39h06a4308_1
ipython=7.29.0=py39hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.0=py39h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py39h06a4308_7
jupyter_client=7.1.0=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libuuid=1.0.3=h7f8727e_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lightfm=1.16=py39hce5d2b2_2
markupsafe=2.0.1=py39h27cfd23_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.3.0=py39h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.6=py39h06a4308_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py39h8c16a72_0
pandocfilters=1.4.3=py39h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.1.3=py39h06a4308_0
pluggy=1.0.0=py39h06a4308_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py=1.10.0=pyhd3eb1b0_0
pycosat=0.6.3=py39h27cfd23_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
pytest=6.2.5=py39h06a4308_2
python=3.9.5=h12debd9_4
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pyzmq=22.3.0=py39h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
ruamel_yaml=0.15.100=py39h27cfd23_0
scikit-learn=1.0.1=py39h51133e4_0
scipy=1.7.1=py39h292c36d_2
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=52.0.0=py39h06a4308_0
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
terminado=0.9.4=py39h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py39h27cfd23_0
tqdm=4.61.2=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pyhd3eb1b0_1
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py39h06a4308_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
